Kernel: Add my own MemoryAllocator implementation
Not sure where the previous malloc implementation was comming from,
maybe it was defined as a weak symbol?

The crash seems to be caused by 'Vector<String>' which seems to call the
destructor of one of the strings twice.  I suspect this has something to
do with moving the string around when reallocating.