Kernel: Add Worker thread that can unblock
This is at a very early state, somehow the thread is not scheduled with
privileges, even though it has m_privileged=true. (We crash because we
need to access some memory which is prohibited without the default
region.)

In the isr_pendsv we return 0xfffffffd does this change the privileges?
I don't think so.