Build: Add 'alloc' bit in '.boot_2_flash_second_stage' section
If I run:

    ./Scripts/build.sh
    ./Scripts/debugger.sh
    > load
    > x/256bx 0x10000000

I can see that the '.boot_2_flash_second_stage' section is now loaded
correctly into memory.  At least it appears that way.  The interesting
thing is that if I disconnect the debugger and then reconnect, I read
different values.  The second time, everything is correct.

The problem is that if I start the application, the flash appears to be
cleared and the data is never copied into memory.

I think what happens is the following:

-   When we do 'load' then 'openocd' will do all sorts of weird things
    which makes it possible to write to flash.

-   When we 'run', the flash is no longer connected with XIP and thus we
    can't see the values anymore, but they are still there.

-   The stuff is never loaded into RAM, because the checksum is wrong.
    If I understand the manual correctly, it should copy things into RAM
    before checking the checksum, but maybe this is an optimization?

What I don't understand is, why I am unable to set a breakpoint in the
boot code.  In my understanding, breakpoints should persist even if I
reset the machine.