Boot: Correctly load the boot loader from GDB
I thought that when we used 'load' from GDB, this would be equivalent to
putting things into flash and pushing reset, but this is not the case.

Instead, it loads everything to the load-memory-address and then jumps
to the entry address.  This means that the boot loader could be executed
at '0x10000000' or '0x20000000' depending on context.

This is really ugly in my opinion and I will do it slightly differently.
Instead, there is a special 'boot_0_openocd_entry' function which is
marked as entry point.  GDB will call this function which relays to
'boot_1_reset', while the reset will trigger it directly.

Therefore, in theory, both contexts would result in the same behaviour
which is much easier to reason about.  However, this doesn't quite work
at the moment, because of some issues with the linker (refer to 0003).