Boot: Load '.boot_0_openocd_entry' in RAM
Now it should theoretically be possible to run the application both with
and without debugger and eventually reach 'boot_2_flash_second_stage'.

However, when we execute with a debugger, we sometimes reach
'boot_0_openocd_entry' and then crash in a random location.

I suspect that we don't quite run the reset handler correctly.  Maybe we
can instead use on of the helper functions in ROM?