Boot: Add 'boot_1_debugger' to emulate 'boot_1_reset'
Instead of manually triggering the reset, we can simply emulate the
behaviour by copying 256 bytes into RAM.

Currently, this does not work.  It appears that 'g++' tries to translate
our code into an unaligned load instruction, which is not supported by
this processor.  Maybe I need to give it some more flags.

The quick solution would be to write my own 'memcpy' implementation from
hand, but that would only work around the issue.