[WIP] Boot: Write 'boot_2_flash_second_stage' in C++
This doesn't quite work, because we loose all the symbol information
when we make the conversion.  I tried to work around this issue, but
it's not trivial.

It seems that the C++ code will generate a 'push' instruction at the
beginning.  I am not sure if this is the particular code that causes the
issue, but we do crash at some point.

I think, I should roll back and implement this in Assembly instead.
Not much needs to happen anyways.