Boot: Implement bootloader entirely in Assembly
After messing around with C++ a bit, I think it is too difficult to use
C++ in this context.

I also decided not to support the entry with 'continue'.  Instead, if we
run from a debugger, we need to use 'run'.

We now reliably enter 'boot_2_flash_second_stage' in RAM.