Boot: Split flash initialization into seperate stage
I started working on the vector tables but realized, that I was missing
the code that initialized the flash for me.  I thought, that we jump to
'_entry_point' from the bootrom, however, this is already in flash.

The next step will be to import the flash logic from the pico-sdk
project.  Then I will have to debug all of this code here.