Boot: Load kernel into memory
Currently, we crash in 'isr_reset' when we try to call
'boot_4_load_kernel'.  This appears to be, because the branch
instruction has the wrong target address, it does not include the thumb
bit.

Additionally, I suspect, that the '_memcpy' function in the bootrom has
a different order of arguments.  I seem to recall having issues with
this.