Build: Link 'linker stubs' into '.text' section
This stuff is generated by the linker for some reason, I am not going to
question it too much.

However, this still doesn't work, because in 'boot_3_vectors' we appear
to call 'boot_4_load_kernel' dynamically instead of statically.  We
really don't want to get a '.plt' involved in the kernel.