Kernel+Drivers: Add 'UartDriver'
I've now encountered two annoying compiler bugs.

-   Namespaces do not work correctly.
    This can be avoided with the 'export import workaround'.

-   We need to import 'kit', because the 'new operator' is defined
    there.  In my understanding, that should work as is.

Both of these issues have workarounds and I can continue, but I have
another issue.  The compiler complains that the 'Singleton<T>::m_initialized'
is not defined anywhere.  I do not know, if this is another compiler bug
or not, I seem to recall writing a workaround in the old version.