Kernel: Resolve ABA problem in UartDriver
The problem wasn't quite an ABA problem but is was about to become one.
Previously, the UART would write into the buffer without checking, which
is obviously bad.  However, fixing this isn't trivial because we need to
be careful that we can distinguish if there are zero bytes produced or
if the buffer is full.

There are so many race conditions in here.  Most of them should be easy
to resolve, but that quickly adds up.