Kernel: Configure UART0 in 'UartDriver'
This seems like a massive challenge, I completely underestimated this.
Maybe I should use the mappings from the SDK, that could save time.

There are tons of details that haven't been configured yet.

All of this is completely untested and I won't be able to test it until
it is completely implemented.