Boot: Hold CORE1 in 'wait_for_vector'
Somehow, this still doesn't work.  We still hit
'UartOutputDriver::UartOutputDriver' twice.

The CORE1 would explain this because it uses the 'memset' again and
can't see the 'm_initialized'.

But for some reason I can't get a hold of the other core, it seems that
GDB and openocd are broken and I can't fix that.