Kernel: Diagnose crash with double stash of default thread
I strongly suspect, that the default thread blocks on the allocator lock
which shouldn't really happen and that suggests that somebody else is
holding the lock but this should be a deadlock and not a crash.

I should add a backup default thread that is scheduled when the default
thread is blocked.