Kernel: Add proper synchronization between thread and handler mode
Previously, I was relatively lazy about this.  Now, I am doing it
properly.

I haven't yet tested if there are any synchronization issues that now
trigger and assertion, or if this helped to resolve the issues I was
having.

There are also a few places left that need some amount of protection.