Kernel: This appears to be a livelock
My understanding is, that the system call worker will loop indefinetely
and therefore, the default thread doesn't get a turn.

In theory, if there is something to clean up, the default thread should
have priority, but that does not appear to be working.

If that were true, this would be a livelock and not a deadlock.