Kernel: Resolve the deadlock issue
The cause was, that I was using a KernelMutex for the debug stuff, that
didn't really make sense and was unnecessary.

In the future, I will need to do this properly, however, that will
require careful considerations, about which actions can occur in
interrupt handlers.

However, I immediatelly ran into the next issue with the page allocator.