Kernel+Synchronization: Add 'HardwareSpinLock'
The goal here is to rework all the locking primitives.  At some point, I
can go through the codebase and actually start using them.  Once all of
that is in place, it should be possible to schedule on multiple cores.

Currently, all of this code is untested and probably doesn't even
compile.